+++ api/src/main/java/org/allaymc/api/permission/DefaultPermissions.java
@@ -27,8 +27,8 @@ public final class DefaultPermissions {
                     .addPermission(DOORS_AND_SWITCHES).addPermission(OPEN_CONTAINERS)
                     .addPermission(ATTACK_PLAYERS).addPermission(ATTACK_MOBS)
                     .addPermission(SUMMON_LIGHTNING).addPermission(PVM).addPermission(MVP)
-                    .addPermission(Command.COMMAND_PERM_PREFIX + "me")
-                    .addPermission(Command.COMMAND_PERM_PREFIX + "version");
+                    .addPermission(Command.COMMAND_PERMISSION_PREFIX + "me")
+                    .addPermission(Command.COMMAND_PERMISSION_PREFIX + "version");
 
     /**
      * The operator permission tree.