+++ CHANGELOG.md
@@ -10,7 +10,11 @@ and any changes to API will have a prefix `(API)`.
 
 Unless otherwise specified, any version comparison below is the comparison of server version, not API version.
 
-## 0.2.0 (API 0.5.0) - Unreleased
+## 0.2.1 (API 0.6.0) - Unreleased
+
+<small>[Compare with 0.2.0](https://github.com/AllayMC/Allay/compare/0.2.0...HEAD)</small>
+
+## 0.2.0 (API 0.5.0) - 2025-3-3
 
 <small>[Compare with 0.1.3](https://github.com/AllayMC/Allay/compare/0.1.3...HEAD)</small>